refactor: kebab-case filenames
`unicorn/filename-case` is on in magic-oxlint-config. Applied with
`magic-kebab --write`: 11 renames, 21 import specifiers rewritten, no
conflicts and nothing needing review.
