+++ src/android/settings-dependency.ts
@@ -1,7 +1,7 @@
 import { ConfigPlugin, withSettingsGradle } from "expo/config-plugins";
 
-import { PluginConfigType } from "../pluginConfig";
-import { codePushGradlePath } from "../utils/codePushGradlePath";
+import { PluginConfigType } from "../plugin-config";
+import { codePushGradlePath } from "../utils/code-push-gradle-path";
 
 export function applySettings(gradleSettings: string) {
   const includeCodePush = "include ':react-native-code-push'";
