+++ src/plugin-config.ts
